+++ pytket/tests/passes_serialisation_test.py
@@ -1,4 +1,4 @@
-# Copyright 2019-2022 Cambridge Quantum Computing
+# Copyright 2019-2023 Cambridge Quantum Computing
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -289,6 +289,7 @@ TWO_WAY_NONPARAM_PASSES = [
     "RemoveBarriers",
     "DecomposeBridges",
     "CnXPairwiseDecomposition",
+    "RemoveImplicitQubitPermutation",
 ]
 
 TWO_WAY_PASSES = {name: nonparam_pass_dict(name) for name in TWO_WAY_NONPARAM_PASSES}