+++ tket/src/Predicates/PassGenerators.cpp
@@ -1,4 +1,4 @@
-// Copyright 2019-2022 Cambridge Quantum Computing
+// Copyright 2019-2023 Cambridge Quantum Computing
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -694,7 +694,6 @@ PassPtr gen_synthesise_pauli_graph(
   Transform t = Transforms::synthesise_pauli_graph(strat, cx_config);
   PredicatePtr ccontrol_pred = std::make_shared<NoClassicalControlPredicate>();
   PredicatePtr mid_pred = std::make_shared<NoMidMeasurePredicate>();
-  PredicatePtr wire_pred = std::make_shared<NoWireSwapsPredicate>();
   OpTypeSet ins = {OpType::Z,       OpType::X,           OpType::Y,
                    OpType::S,       OpType::Sdg,         OpType::V,
                    OpType::Vdg,     OpType::H,           OpType::CX,
@@ -708,7 +707,6 @@ PassPtr gen_synthesise_pauli_graph(
   PredicatePtrMap precons{
       CompilationUnit::make_type_pair(ccontrol_pred),
       CompilationUnit::make_type_pair(mid_pred),
-      CompilationUnit::make_type_pair(wire_pred),
       CompilationUnit::make_type_pair(in_gates)};
   PredicateClassGuarantees g_postcons = {
       {typeid(ConnectivityPredicate), Guarantee::Clear},